requirements.txt
spacy>=2.1.3
numpy>=1.15.4
gensim>=3.7.2
pandas>=0.24.2
nltk >= 3.4.1
newspaper3k >= 0.2.8
urlextract >= 0.13.0
tqdm >= 4.33.0